Move diff header to diff directory
